main/requirements.txt
discord.py == 1.7.3
googletrans == 4.0.0rc1